Initial port of delegate resources docs
This is an initial port for the delegate resources documentation from
the DP0.2 site:

- https://dp0-2.lsst.io/dp0-delegate-resources/index.html
- https://github.com/lsst/dp0-2_lsst_io/tree/main/dp0-delegate-resources

Some of the cross-references can't be made either because the doc hasn't
been ported yet (e.g. tutorials) or because the right docs are ported to
rsp.lsst.io yet. I've commented these out and added a TODO comment to
indicate what needs to be re-added.